Move ScanPackage message
