+++ src/Controller/OrganizationController.php
@@ -19,12 +19,12 @@ use Buddy\Repman\Message\Organization\Package\AddGitLabHook;
 use Buddy\Repman\Message\Organization\Package\RemoveBitbucketHook;
 use Buddy\Repman\Message\Organization\Package\RemoveGitHubHook;
 use Buddy\Repman\Message\Organization\Package\RemoveGitLabHook;
-use Buddy\Repman\Message\Organization\Package\ScanPackage;
 use Buddy\Repman\Message\Organization\RegenerateToken;
 use Buddy\Repman\Message\Organization\RemoveOrganization;
 use Buddy\Repman\Message\Organization\RemovePackage;
 use Buddy\Repman\Message\Organization\RemoveToken;
 use Buddy\Repman\Message\Organization\SynchronizePackage;
+use Buddy\Repman\Message\Security\ScanPackage;
 use Buddy\Repman\Query\User\Model\Installs\Day;
 use Buddy\Repman\Query\User\Model\Organization;
 use Buddy\Repman\Query\User\Model\Package;
